move workers to separate files - added infrared
