+++ src/Utils.h
@@ -1,0 +1,15 @@
+#ifndef Kinect2_Utils_h
+#define Kinect2_Utils_h	
+
+// Safe release for interfaces
+template<class Interface>
+inline void SafeRelease(Interface *& pInterfaceToRelease)
+{
+	if (pInterfaceToRelease != NULL)
+	{
+		pInterfaceToRelease->Release();
+		pInterfaceToRelease = NULL;
+	}
+}
+
+#endif
